fix: clean up and questions
