+++ packages/ui/components/Button/index.tsx
@@ -4,10 +4,19 @@
 
 import React from 'react';
 import styled, { css } from '@refract-ui/sc';
+import type { DefaultTheme } from 'styled-components';
 
-import type { PayloadButtonProps } from '../../primitives/PayloadButton';
+type colorProps = keyof DefaultTheme['themeColors'];
 
-const StyledButton = styled.button<{ color: PayloadButtonProps['color'] }>`
+export type PayloadButtonProps = {
+  color: colorProps;
+  size?: 'sm' | 'md' | 'lg';
+  type?: 'button' | 'reset' | 'submit';
+  variant?: 'solid' | 'outlined' | 'link';
+  onClick?: () => void;
+};
+
+const StyledButton = styled.button<PayloadButtonProps>`
   ${({ theme: { themeColors }, color }) => css`
     background-color: ${themeColors[color]};
   `}
@@ -16,11 +25,19 @@ const StyledButton = styled.button<{ color: PayloadButtonProps['color'] }>`
 function Button({
   type = 'button',
   color = 'primary',
+  size = 'md',
+  variant = 'solid',
   onClick = () => console.log('@--> click')
 }: PayloadButtonProps) {
   return (
     // eslint-disable-next-line react/button-has-type
-    <StyledButton onClick={onClick} type={type} color={color}>
+    <StyledButton
+      onClick={onClick}
+      type={type}
+      color={color}
+      size={size}
+      variant={variant}
+    >
       Boop
     </StyledButton>
   );
