+++ packages/ui/primitives/PayloadText.ts
@@ -1,0 +1,1 @@
+export type PayloadTextProps = string | number | null;
